Move base and derived sensors to entity.py
